Organizes test files
