+++ tests/assertions/test_ssl_on.rb
@@ -1,4 +1,4 @@
-ssl_on_config_file = 'tests/test_ssl_on.yaml'
+ssl_on_config_file = 'tests/configs/test_ssl_on.yaml'
 ssl_on_dns_hosts_file = File.join(File.dirname(__FILE__), 'test_ssl_on_dns_hosts.txt')
 ssl_on = Config.new(ssl_on_config_file, ssl_on_dns_hosts_file)
