+++ tests/assertions/test_utilities.rb
